+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.4.0] - 2025-04-01
+### Added / Changed / Fixed
+- Changed: function drag_mouse to drag_mouse_to_pos. Reason is the new drag_mouse function that moves mouse in relation to current position
+- Added functions: 
+-    - move_mouse() - Moves mouse relative to its current position. -x left, +x right, -y up, +y down
+-    - move_mouse_to() -- Accepts Option<x>, Option<y>. None value keeps the same x or y. Usefull for horizontal and vertical movement
+-    - drag_mouse() - performs mouse drag action relative to current position. Same rules as in move_mouse function
+-    - drag_mouse_to() - same as in mouse_move_to, accepts Options. 
+-    - get_mouse_position() - returns Result<(i32, i32)> current mouse position
+- move_mouse_to_pos() remains the same, while drag_mouse_to_pos() is new name for the old version of drag_mouse() function
+
 ## [2.3.0] - 2025-03-30
 ### Fixed / Removed
 - Rework of Segmented NCC template match. Completely removed argument of Max segments and made it always work robustly, never sacrificing precision for  speed. Additionally fixed part of formulas which will additionally reduce false positives, regardless of max segments. The fix also improves algorithms speed when compared to previous version, if max_segments is not taken into consideration. The speed gain is due to much less checks and verifications in the algorithm.